+++ internal/controllers/common/resource_definitions/imagetag_generated.go
