+++ internal/controllers/recording_controller_test.go
@@ -52,9 +52,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	operatorv1beta1 "github.com/cryostatio/cryostat-operator/api/v1beta1"
-	"github.com/cryostatio/cryostat-operator/controllers"
-	cryostatClient "github.com/cryostatio/cryostat-operator/controllers/client"
-	"github.com/cryostatio/cryostat-operator/test"
+	"github.com/cryostatio/cryostat-operator/internal/controllers"
+	cryostatClient "github.com/cryostatio/cryostat-operator/internal/controllers/client"
+	"github.com/cryostatio/cryostat-operator/internal/test"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 )
